refactor: rename macro 'entry_def' to 'entry_type'
