+++ crates/hdk_derive/src/entry_types_name_registration.rs
